+++ SystemInventorys/SystemInventory-Service/src/main/java/zhongchiedu/inventory/service/StockService.java
@@ -56,7 +56,7 @@ public interface StockService extends GeneralService<Stock> {
 	
 	public Set<String> findProjectNames();
 	
-	public void preStockToStock(PreStock preStock,long actnum);
+	public void preStockToStock(PreStock preStock,Double actnum);
 	
 	public BasicDataResult pickUpApplicationToStock(PickUpApplication pickUpApplication);
